apple_mobility_etl/requirements.txt
attrs==20.1.0
coverage==5.2.1
iniconfig==1.0.1
more-itertools==8.5.0
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
